Add small flex example
flex is an open source implementation of lex.
lex is a tool that takes an input of a language definition and creates
a lexer in C for that language.
This example creates a lexer for a simple config file.
The lexer function of myscanner.c utilizes this lexer.
The main.go file calls the lexer function using cgo.