Start working on actual project
Right now it looks like it will be best to have a token type in the analyzer class, and have a lexer function call a subpackage lexer based on the language being used.
I've added a cpp folder for the c++ lexer stuff.
It contains the scanner.h/.c, lex.yy.c, and the go file to use them.